Transition between levels, fixed touchscreenkeyboard issue
Creates a nicer transition between levels, as it was kind of jarring before. Also there was an issue with the keyboard where when you open the code editor on a mobile device, the keyboard would pop up but none of the text you'd type actually showed up in the editor. Just prevented the keyboard from showing up, saving some frustration of writing code only to realize it didn't actually show in the editor (have to tap on the editor to open keyboard though)